requirements.txt
jinja2
PyYAML
snakemake
